Feature 06: Documented and Refactored

+++ src/main/java/krisbanas/awsdemo/services/BasicResourceService.java
@@ -1,11 +1,17 @@
 package krisbanas.awsdemo.services;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.*;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
+import com.amazonaws.services.s3.model.GetObjectRequest;
+import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.transfer.TransferManager;
 import com.amazonaws.services.s3.transfer.TransferManagerBuilder;
 import com.amazonaws.util.IOUtils;
+import krisbanas.awsdemo.exceptions.BucketException;
 import krisbanas.awsdemo.exceptions.ResourceExceptionException;
+import krisbanas.awsdemo.services.interfaces.ResourceService;
+import krisbanas.awsdemo.services.utils.FileUtils;
+import krisbanas.awsdemo.services.utils.S3Utils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
@@ -14,49 +20,43 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Objects;
-import java.util.UUID;
 
 @Slf4j
 @Service
-public class AmazonS3Service {
+public class BasicResourceService implements ResourceService {
 
     private final AmazonS3 amazonS3;
     private final TransferManager transferManager;
 
     @Autowired
-    public AmazonS3Service(AmazonS3 amazonS3) {
+    public BasicResourceService(AmazonS3 amazonS3) {
         this.amazonS3 = amazonS3;
         transferManager = TransferManagerBuilder.standard()
                 .withS3Client(amazonS3)
                 .build();
     }
 
+    @Override
     public String storeResourceToBucket(MultipartFile file, String bucketName) throws IOException {
-        validateFile(file, bucketName);
-        var request = preparePutObjectRequest(file, bucketName);
+        if (!amazonS3.doesBucketExistV2(bucketName)) throw new BucketException("Bucket does not exist");
+        FileUtils.validateFile(file, bucketName);
+        var request = S3Utils.preparePutObjectRequest(file, bucketName);
         amazonS3.putObject(request);
         log.info("Stored new file with id {} in bucket {}", request.getKey(), bucketName);
         return request.getKey();
     }
 
+    @Override
     public String storeResourceStreamed(MultipartFile file, String bucketName) throws IOException, InterruptedException {
-        validateFile(file, bucketName);
+        if (!amazonS3.doesBucketExistV2(bucketName)) throw new BucketException("Bucket does not exist");
+        FileUtils.validateFile(file, bucketName);
         String key = storeFileAndWait(file, bucketName);
         log.info("Stored new file with id {} in bucket {}", key, bucketName);
         return key;
     }
 
-    private String storeFileAndWait(MultipartFile file, String bucketName) throws IOException, InterruptedException {
-        var key = generateResourceKey();
-        File newFile = convertMultipartFileToFile(file);
-        var upload = transferManager.upload(bucketName, key, newFile);
-        upload.waitForCompletion();
-        return key;
-    }
-
+    @Override
     public Resource getResource(String bucketName, String resourceId) throws IOException {
         var getObjectRequest = new GetObjectRequest(bucketName, resourceId);
         var fullObject = tryFetchingResource(getObjectRequest);
@@ -65,6 +65,7 @@ public class AmazonS3Service {
         return new ByteArrayResource(byteArray);
     }
 
+    @Override
     public boolean deleteResource(String bucketName, String resourceId) {
         try {
             amazonS3.getObject(bucketName, resourceId);
@@ -75,46 +76,20 @@ public class AmazonS3Service {
         return true;
     }
 
+    private String storeFileAndWait(MultipartFile file, String bucketName) throws IOException, InterruptedException {
+        var key = S3Utils.generateResourceKey();
+        File newFile = FileUtils.convertMultipartFileToFile(file);
+        var upload = transferManager.upload(bucketName, key, newFile);
+        upload.waitForCompletion();
+        newFile.delete();
+        return key;
+    }
+
     private S3Object tryFetchingResource(GetObjectRequest getObjectRequest) {
         try {
             return amazonS3.getObject(getObjectRequest);
         } catch (AmazonS3Exception ex) {
             throw new ResourceExceptionException("No resource found");
         }
-    }
-
-    private File convertMultipartFileToFile(MultipartFile file) throws IOException {
-        var newFile = new File(Objects.requireNonNull(file.getOriginalFilename()));
-        file.getInputStream().transferTo(new FileOutputStream(newFile));
-        return newFile;
-    }
-
-    private PutObjectRequest preparePutObjectRequest(MultipartFile multipartFile, String bucketName) throws IOException {
-        var inputStream = multipartFile.getInputStream();
-        var key = generateResourceKey();
-        ObjectMetadata metadata = prepareMetadata();
-        return new PutObjectRequest(bucketName, key, inputStream, metadata);
-    }
-
-    private void validateFile(MultipartFile multipartFile, String bucketName) {
-        if (multipartFile == null) {
-            log.warn("No file to store in bucket {}", bucketName);
-            throw new ResourceExceptionException("No file to store sent");
-        }
-        if (multipartFile.isEmpty()) {
-            log.warn("The file to store is empty. Ignoring...");
-            throw new ResourceExceptionException("Empty file sent");
-        }
-    }
-
-    private ObjectMetadata prepareMetadata() {
-        var metadata = new ObjectMetadata();
-        metadata.setContentType("plain/text");
-        metadata.addUserMetadata("metadataKey", "metadataValue");
-        return metadata;
-    }
-
-    private String generateResourceKey() {
-        return UUID.randomUUID().toString();
     }
 }
